requirements.txt
astroid==1.6.5
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-bootstrap3==10.0.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.1.0
pkg-resources==0.0.0
pycodestyle==2.4.0
pylint==1.9.2
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
